requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==4.4.2
distro==1.9.0
elevenlabs==1.50.3
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.6
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipython==8.31.0
JayDeBeApi==1.2.3
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jpype1==1.5.1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==0.28.1
openai-whisper==20240930
packaging==24.2
parso==0.8.4
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2-binary==2.9.10
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.37
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
